pyproject.toml
[project]
name = "comfyui-video-frame-extractor"
version = "1.2.0"
description = "ComfyUI custom node that brings a DAW-style interactive video timeline directly into the node graph. Upload any video, scrub through it in real time, drag a loop region to define your extraction window, and pipe the resulting frame batch into any downstream node."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["comfyui", "video", "frame-extraction", "mp4", "custom-node"]
license = { text = "MIT" }
dependencies = [
    "opencv-python>=4.7",
    "Pillow>=9.0",
    "numpy>=1.22",
    "torch",
]

[project.urls]
Homepage = "https://github.com/comfyuiattic-989/ComfyUI-Video-Frame-Extractor"
Repository = "https://github.com/comfyuiattic-989/ComfyUI-Video-Frame-Extractor"

[tool.comfy]
PublisherId = "comfyui-attic"
DisplayName = "ComfyUI-Video-Frame-Extractor"
Icon = "https://avatars.githubusercontent.com/u/273807868?v=4&size=512"
